requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
faiss-cpu
google-generativeai>=0.3.0
httpx
python-dotenv
fastapi-users[sqlalchemy]
passlib[bcrypt]
aiosqlite
requests
transformers
sentencepiece
torch
greenlet
argon2-cffi
celery
redis
